requirements.txt
aiohttp==3.8.6
tweepy==4.14.0
python-telegram-bot==20.7
APScheduler==3.10.4
pytz==2023.3
requests==2.31.0
openai
